requirements.txt
torch
torchvision>=0.5
opencv-python-headless<=4.5.4.60
scipy
numpy
Pillow
scikit-image
python-bidi
PyYAML
Shapely
pyclipper
ninja
absl-py==1.4.0
appdirs==1.4.4
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
attr==0.3.1
attrs==22.2.0
azure-core==1.26.3
azure-storage-blob==12.14.1
bleach==5.0.1
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
boxing==0.1.4
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
cycler==0.11.0
defusedxml==0.7.1
Django==3.2.16
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-swagger==2.2.0
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
easydict==1.10
expiringdict==1.2.2
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-api-core==2.10.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.3
grpcio==1.51.1
grpcio-status==1.48.2
h5py==3.8.0
htmlmin==0.1.12
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==3.2.0
keras==2.11.0
kiwisolver==1.4.4
label-studio==1.7.1
label-studio-converter==0.0.48rc0
label-studio-tools==0.0.1
launchdarkly-server-sdk==7.5.0
libclang==15.0.6.1
lockfile==0.12.2
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
msrest==0.7.1
nltk==3.6.7
numpy==1.24.1
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==23.0
pandas==1.3.5
Pillow==9.3.0
proto-plus==1.22.2
protobuf==3.20.3
psycopg==3.1.8
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pyparsing==3.0.9
pypi==2.1
pypl==40.10
pyRFC3339==1.1
pyrsistent==0.19.3
python-dateutil==2.8.1
python-json-logger==2.0.4
pytz==2019.3
PyYAML==6.0
redis==4.4.2
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
rq==1.10.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
rules==2.2
s3transfer==0.3.7
scikit-learn==1.2.1
scipy==1.10.1
semver==2.13.0
sentry-sdk==1.14.0
simplejson==3.18.3
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
svgwrite==1.4.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-macos==2.11.0
termcolor==2.2.0
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.4.0
ua-parser==0.16.1
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.14
user-agents==2.2.0
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
xmljson==0.2.0
zipp==3.12.0